0.1.9.1 Update Release
MoreSkills 0.1.9 NEW!

________________________________________

New Jump Overhaul:

1- Now you can jump higher if you press Shift (Optional and custom if wanted).

a. Can change the base jump force, and the max when level 100 is reached.
b. Can change the Custom Keybind to activate the higher jump or deactivate this key.
c. Can disable Higher Jump if you don't want it.
d. Have a Decreased fall damage, with custom divider, reached at level 100.
e. Added a roll if you fall from heights from 4 to 8 meters(?). In the future this roll will have decreased fall damage and stamina with Vitality and Strength levels respectly.

2- Revamped Crafting, Hunting and Taming.

a. Now Simple Recycling is attached to your Crafting level, so the higher the cost the higher the recycling, but if you made something that cost 100 wood and then when you are level 100 it will read the cost then and give you back the selected amount (default 50%).
b. Now it will get and save all new creatures and items. In case of hunting it will get what it drops and do the math.
c. Now the drops take in count the creature's level and multiply it.
d. Now Taming will add al custom Creatures, and if not in the list it will add Default Taming Levels, Unlock: 30, Tamer: 60, Master: 90.

3- Fixed Bugs

a. Fixed bug in WoodCutting where FineWood would not drop.
b. Fixed bug in Taming that when looking to a boar it would send a log error.
c. Fixed the Wheel Scroll Killer Config.
d. Other minor fixes.

________________________________________________________